restructured file layout
currently working on restructured file layout to be more inline with
standard. Added cmake to workflow to build project. Currently the
CMakeLists are not correct yet.
