+++ src/TUI_components/include_TUI_components/components.h
@@ -1,0 +1,10 @@
+#ifndef COMPONENTS_H   /* Include guard */
+#define COMPONENTS_H
+
+// aggregator file so I dont need to change all the inports in every file
+
+#include "hours.h"
+#include "days.h"
+#include "week.h"
+
+#endif
