+++ src/TUI_components/src/hours.c
@@ -1,7 +1,7 @@
 
 #include <curses.h>
 #include <ncurses.h>
-#include "hours.h"
+#include "../include_TUI_components../hours.h"
 #include "customutils.h"
 
 struct Hours hoursConst(int hourLength, int hourYOffset,  WINDOW *pw) {
