+++ src/customutils/src/customutils.c
@@ -1,6 +1,6 @@
 #include <ncurses.h>
 #include <curses.h>
-#include "customutils.h"
+#include "../include_customutils/customutils.h"
 
 int max(int a, int b) {
   return (a>=b)*a + (b>a)*b;
